Refactor mcp-service.ts into focused modules
Split monolithic 2,722-line mcp-service.ts into 10 focused modules:

1. **ServerLifecycleManager.ts** (~450 lines)
   - Server initialization and connection
   - Server shutdown and cleanup
   - Server restart
   - Client request handlers (elicitation, sampling)
   - Server status tracking

2. **TransportFactory.ts** (~210 lines)
   - Transport creation (stdio, WebSocket, HTTP)
   - Command path resolution
   - Environment variable preparation

3. **OAuthManager.ts** (~270 lines)
   - OAuth client management
   - OAuth flows (initiate, complete)
   - Token storage and retrieval
   - OAuth status checking
   - Token revocation

4. **ToolManager.ts** (~280 lines)
   - Tool filtering and enable/disable state
   - Profile-based tool filtering
   - Detailed tool list generation
   - Orphaned tool cleanup

5. **ToolExecutor.ts** (~485 lines)
   - Tool call execution
   - Parameter type fixing and normalization
   - Retry logic with corrected parameters
   - Tool approval flow
   - Session-aware tool filtering

6. **ResponseProcessor.ts** (~200 lines)
   - Response filtering to reduce context size
   - Large response summarization
   - Chunking and processing strategies

7. **ServerStateManager.ts** (~235 lines)
   - Runtime enabled/disabled server tracking
   - Profile MCP configuration application
   - Server availability checking
   - Initialization state management

8. **ResourceTracker.ts** (~135 lines)
   - Resource lifecycle tracking (sessions, connections)
   - Resource activity updates
   - Inactive resource cleanup
   - Resource ID extraction from tool results

9. **ServerLogger.ts** (~65 lines)
   - Server log storage with circular buffer
   - Log retrieval and clearing

10. **index.ts** (~595 lines)
    - MCPService facade that composes all modules
    - Unified API surface for backward compatibility

11. **types.ts** (~25 lines)
    - Shared type definitions

**Benefits:**
- Improved maintainability through separation of concerns
- Easier testing of individual components
- Better code navigation and understanding
- Preserved all existing functionality
- Maintained backward compatibility via re-exports

**Total line reduction:** 2,722 lines → ~2,950 lines across 11 files
(slight increase due to module boundaries and documentation)
